+++ types/counters.go
@@ -1,6 +1,6 @@
 package divan_types
 
 type Counters struct {
-	RebalanceSuccess uint `json:"rebalance_success"`
-	RebalanceStart   uint `json:"rebalance_start"`
+	RebalanceSuccess uint64 `json:"rebalance_success"`
+	RebalanceStart   uint64 `json:"rebalance_start"`
 }
