+++ types/node.go
@@ -5,11 +5,11 @@ type Node struct {
 	SystemStats          SystemStats        `json:"systemStats"`
 	InterestingStats     InterestingStats   `json:"interestingStats"`
 	Uptime               string             `json:"uptime"`
-	MemoryTotal          uint               `json:"memoryTotal"`
-	MemoryFree           uint               `json:"memoryFree"`
-	McdMemoryReserved    uint               `json:"mcdMemoryReserved"`
-	McdMemoryAllocated   uint               `json:"mcdMemoryAllocated"`
-	Replication          uint               `json:"replication"`
+	MemoryTotal          uint64               `json:"memoryTotal"`
+	MemoryFree           uint64               `json:"memoryFree"`
+	McdMemoryReserved    uint64               `json:"mcdMemoryReserved"`
+	McdMemoryAllocated   uint64               `json:"mcdMemoryAllocated"`
+	Replication          uint64               `json:"replication"`
 	ClusterMembership    string             `json:"clusterMembership"`
 	RecoveryType         string             `json:"recoveryType"`
 	Status               string             `json:"status"`
@@ -20,7 +20,7 @@ type Node struct {
 	ClusterCompatibility string             `json:"clusterCompatibility"`
 	Version              string             `json:"version"`
 	Os                   string             `json:"os"`
-	CpuCount             uint               `json:"cpuCount"`
+	CpuCount             uint64               `json:"cpuCount"`
 	Ports                Ports              `json:"ports"`
 	Services             []string           `json:"services"`
 	NodeEncryption       bool               `json:"nodeEncryption"`
